Move spyder.py to app namespace
