+++ spyderlib/app/spyder.py
